requirements.txt
python-telegram-bot==13.7
python-dotenv==1.0.0
Pillow==10.0.0
urllib3==1.26.15
pysocks==1.7.1 